requirements.txt
slack-sdk
python-dotenv
tqdm
chromadb==0.5.5
sentence-transformers